+++ core/model/user.go
@@ -20,6 +20,8 @@ const (
 	TypeAccountRoles logutils.MessageDataType = "account roles"
 	//TypeMFAType mfa type
 	TypeMFAType logutils.MessageDataType = "mfa type"
+	//TypeAccountGroups account groups
+	TypeAccountGroups logutils.MessageDataType = "account groups"
 	//TypeProfile profile
 	TypeProfile logutils.MessageDataType = "profile"
 	//TypeDevice device
@@ -37,8 +39,8 @@ type Account struct {
 	AppOrg ApplicationOrganization
 
 	Permissions []Permission
-	Roles       []AppOrgRole
-	Groups      []AppOrgGroup
+	Roles       []AccountRole
+	Groups      []AccountGroup
 
 	AuthTypes []AccountAuthType
 
@@ -54,6 +56,19 @@ type Account struct {
 
 	DateCreated time.Time
 	DateUpdated *time.Time
+}
+
+//GetAccountAuthTypeByID finds account auth type by id
+func (a Account) GetAccountAuthTypeByID(ID string) *AccountAuthType {
+	var result AccountAuthType
+	for _, aat := range a.AuthTypes {
+		if aat.ID == ID {
+			result = aat
+		}
+	}
+	//assign account
+	result.Account = a
+	return &result
 }
 
 //GetAccountAuthType finds account auth type
@@ -100,17 +115,21 @@ func (a Account) GetPermissionsMap() map[string]Permission {
 		permissionsMap[permission.Name] = permission
 	}
 	for _, role := range a.Roles {
-		for _, permission := range role.Permissions {
-			permissionsMap[permission.Name] = permission
+		if role.Active {
+			for _, permission := range role.Role.Permissions {
+				permissionsMap[permission.Name] = permission
+			}
 		}
 	}
 	for _, group := range a.Groups {
-		for _, permission := range group.Permissions {
-			permissionsMap[permission.Name] = permission
-		}
-		for _, role := range group.Roles {
-			for _, permission := range role.Permissions {
+		if group.Active {
+			for _, permission := range group.Group.Permissions {
 				permissionsMap[permission.Name] = permission
+			}
+			for _, role := range group.Group.Roles {
+				for _, permission := range role.Permissions {
+					permissionsMap[permission.Name] = permission
+				}
 			}
 		}
 	}
@@ -126,6 +145,100 @@ func (a Account) GetVerifiedMFATypes() []MFAType {
 		}
 	}
 	return mfaTypes
+}
+
+//GetPermission returns the permission for an ID if the account has it
+func (a Account) GetPermission(id string) *Permission {
+	for _, permission := range a.Permissions {
+		if permission.ID == id {
+			return &permission
+		}
+	}
+	return nil
+}
+
+//GetPermissionNamed returns the permission for a name if the account has it
+func (a Account) GetPermissionNamed(name string) *Permission {
+	for _, permission := range a.Permissions {
+		if permission.Name == name {
+			return &permission
+		}
+	}
+	return nil
+}
+
+//GetActiveRoles returns all active roles
+func (a Account) GetActiveRoles() []AccountRole {
+	roles := []AccountRole{}
+	for _, role := range a.Roles {
+		if role.Active {
+			roles = append(roles, role)
+		}
+	}
+	return roles
+}
+
+//GetRole returns the role for an id if the account has it
+func (a Account) GetRole(id string) *AccountRole {
+	for _, role := range a.Roles {
+		if role.Role.ID == id {
+			return &role
+		}
+	}
+	return nil
+}
+
+//GetActiveGroups returns all active groups
+func (a Account) GetActiveGroups() []AccountGroup {
+	groups := []AccountGroup{}
+	for _, group := range a.Groups {
+		if group.Active {
+			groups = append(groups, group)
+		}
+	}
+	return groups
+}
+
+//GetGroup returns the group for an id if the account has it
+func (a Account) GetGroup(id string) *AccountGroup {
+	for _, group := range a.Groups {
+		if group.Group.ID == id {
+			return &group
+		}
+	}
+	return nil
+}
+
+//AccountRole represents a role assigned to an account
+type AccountRole struct {
+	Role     AppOrgRole
+	Active   bool
+	AdminSet bool
+}
+
+//AccountRolesFromAppOrgRoles converts AppOrgRoles to AccountRoles
+func AccountRolesFromAppOrgRoles(items []AppOrgRole, active bool, adminSet bool) []AccountRole {
+	accountRoles := make([]AccountRole, len(items))
+	for i, role := range items {
+		accountRoles[i] = AccountRole{Role: role, Active: active, AdminSet: adminSet}
+	}
+	return accountRoles
+}
+
+//AccountGroup represents a group assigned to an account
+type AccountGroup struct {
+	Group    AppOrgGroup
+	Active   bool
+	AdminSet bool
+}
+
+//AccountGroupsFromAppOrgGroups converts AppOrgGroups to AccountGroups
+func AccountGroupsFromAppOrgGroups(items []AppOrgGroup, active bool, adminSet bool) []AccountGroup {
+	accountGroups := make([]AccountGroup, len(items))
+	for i, group := range items {
+		accountGroups[i] = AccountGroup{Group: group, Active: active, AdminSet: adminSet}
+	}
+	return accountGroups
 }
 
 //AccountAuthType represents account auth type
@@ -217,6 +330,7 @@ type ExternalSystemUser struct {
 	MiddleName string   `json:"middle_name" bson:"middle_name"`
 	LastName   string   `json:"last_name" bson:"last_name"`
 	Email      string   `json:"email" bson:"email"`
+	Roles      []string `json:"roles" bson:"roles"`
 	Groups     []string `json:"groups" bson:"groups"`
 
 	//here are the system specific data for the user - uiucedu_uin etc
@@ -238,6 +352,9 @@ func (esu ExternalSystemUser) Equals(other ExternalSystemUser) bool {
 		return false
 	}
 	if esu.Email != other.Email {
+		return false
+	}
+	if !utils.DeepEqual(esu.Roles, other.Roles) {
 		return false
 	}
 	if !utils.DeepEqual(esu.Groups, other.Groups) {