+++ utils/utils.go
@@ -6,7 +6,10 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"math/rand"
+	"net/http"
 	"reflect"
+
+	"github.com/rokwire/logging-library-go/logs"
 
 	"github.com/rokwire/logging-library-go/errors"
 )
@@ -81,4 +84,16 @@ func SetStringIfEmpty(a, b string) string {
 //GetType returns a string representing the type of data
 func GetType(data interface{}) string {
 	return reflect.TypeOf(data).String()
+}
+
+//GetIP extracts the IP address from the http request
+func GetIP(l *logs.Log, r *http.Request) string {
+	IPAddress := r.Header.Get("X-Real-Ip")
+	if IPAddress == "" {
+		IPAddress = r.Header.Get("X-Forwarded-For")
+	}
+	if IPAddress == "" {
+		IPAddress = r.RemoteAddr
+	}
+	return IPAddress
 }